refactor(native): isolate tcc toolchain policy
